# openapi.yaml
openapi: 3.0.0
info:
  title: Logistic Backend API
  description: A comprehensive logistic management system API built with NestJS
  version: 1.0.0
  contact:
    name: Logistic Team
    url: https://logistic-backend.com
    email: support@logistic-backend.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.logistic-backend.com
    description: Production server
...
paths:
...
  /api/v1/users:
    post:
      tags:
        - Users
      summary: Create a new user
      description: Create a new user account (Admin only)
      operationId: createUser
      security:
        - JWT-auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: User created successfully
                data:
                  id: uuid
                  email: user@example.com
                  firstName: John
                  lastName: Doe
                  role: user
                  status: active
                  createdAt: '2023-01-01T00:00:00.000Z'
        '400':
          description: Validation error or user already exists
        '403':
          description: Forbidden - Admin role required
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a paginated list of all users (Admin/Manager only)
      operationId: findAllUsers
      security:
        - JWT-auth: []
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: search
          in: query
          description: Search term to filter results
          required: false
          schema:
            type: string
            example: john
        - name: sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            example: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum:
              - ASC
              - DESC
            default: DESC
            example: DESC
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: Users retrieved successfully
                data:
                  - id: uuid
                    email: user@example.com
                    firstName: John
                    lastName: Doe
                    role: user
                    status: active
                meta:
                  page: 1
                  limit: 10
                  total: 1
                  totalPages: 1
        '403':
          description: Forbidden - Admin or Manager role required
  /api/v1/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user by their ID (Admin/Manager only)
      operationId: findOneUser
      security:
        - JWT-auth: []
      parameters:
        - name: id
          in: path
          description: User UUID
          required: true
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: User retrieved successfully
                data:
                  id: uuid
                  email: user@example.com
                  firstName: John
                  lastName: Doe
                  role: user
                  status: active
                  phone: +1234567890
                  createdAt: '2023-01-01T00:00:00.000Z'
                  updatedAt: '2023-01-01T00:00:00.000Z'
        '404':
          description: User not found
        '403':
          description: Forbidden - Admin or Manager role required
    patch:
      tags:
        - Users
      summary: Update user
      description: Update user information (Admin only)
      operationId: updateUser
      security:
        - JWT-auth: []
      parameters:
        - name: id
          in: path
          description: User UUID
          required: true
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: User updated successfully
                data:
                  id: uuid
                  email: user@example.com
                  firstName: John
                  lastName: Doe
                  role: user
                  status: active
                  updatedAt: '2023-01-01T00:00:00.000Z'
        '404':
          description: User not found
        '403':
          description: Forbidden - Admin role required
    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user account (Admin only)
      operationId: removeUser
      security:
        - JWT-auth: []
      parameters:
        - name: id
          in: path
          description: User UUID
          required: true
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: User deleted successfully
        '404':
          description: User not found
        '403':
          description: Forbidden - Admin role required
  /api/v1/users/{id}/change-password:
    patch:
      tags:
        - Users
      summary: Change user password
      description: Change password for a specific user (Admin only)
      operationId: changeUserPassword
      security:
        - JWT-auth: []
      parameters:
        - name: id
          in: path
          description: User UUID
          required: true
          schema:
            type: string
            format: uuid
            example: 123e4567-e89b-12d3-a456-426614174000
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDto'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: Password changed successfully
        '400':
          description: Invalid current password
        '404':
          description: User not found
        '403':
          description: Forbidden - Admin role required
  /api/v1/users/profile/me:
    get:
      tags:
        - Users
      summary: Get my profile
      description: Get current authenticated user profile
      operationId: getMyProfile
      security:
        - JWT-auth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: Profile retrieved successfully
                data:
                  id: uuid
                  email: user@example.com
                  firstName: John
                  lastName: Doe
                  role: user
                  status: active
                  phone: +1234567890
                  createdAt: '2023-01-01T00:00:00.000Z'
                  updatedAt: '2023-01-01T00:00:00.000Z'
        '401':
          description: Unauthorized - Invalid or missing token
    patch:
      tags:
        - Users
      summary: Update my profile
      description: Update current authenticated user profile (role and status cannot be changed)
      operationId: updateMyProfile
      security:
        - JWT-auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: Profile updated successfully
                data:
                  id: uuid
                  email: user@example.com
                  firstName: John
                  lastName: Doe
                  role: user
                  status: active
                  updatedAt: '2023-01-01T00:00:00.000Z'
        '400':
          description: Validation error
        '401':
          description: Unauthorized - Invalid or missing token
  /api/v1/users/profile/change-password:
    patch:
      tags:
        - Users
      summary: Change my password
      description: Change password for current authenticated user
      operationId: changeMyPassword
      security:
        - JWT-auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordDto'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseDto'
              example:
                success: true
                message: Password changed successfully
        '400':
          description: Invalid current password
        '401':
          description: Unauthorized - Invalid or missing token
...
components:
  securitySchemes:
    JWT-auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token
  schemas:
    ApiResponseDto:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
          example: true
        message:
          type: string
          description: Response message
          example: Operation completed successfully
        data:
          description: Response data payload
        meta:
          type: object
          description: Pagination metadata
          properties:
            page:
              type: number
              example: 1
            limit:
              type: number
              example: 10
            total:
              type: number
              example: 100
            totalPages:
              type: number
              example: 10
        errors:
          type: array
          items:
            type: string
          description: Array of error messages
          example:
            - Validation failed
            - Invalid input
...
    CreateUserDto:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        password:
          type: string
          minLength: 6
          description: User password
          example: password123
        firstName:
          type: string
          maxLength: 50
          description: User first name
          example: John
        lastName:
          type: string
          maxLength: 50
          description: User last name
          example: Doe
        phone:
          type: string
          description: User phone number
          example: +1234567890
        role:
          type: string
          enum:
            - admin
            - user
            - driver
            - manager
          description: User role
          example: user
        address:
          type: string
          description: User address
          example: 123 Main St
        city:
          type: string
          description: User city
          example: New York
        country:
          type: string
          description: User country
          example: United States
        postalCode:
          type: string
          description: User postal code
          example: 10001
        dateOfBirth:
          type: string
          format: date
          description: User date of birth
          example: '1990-01-01'
    UpdateUserDto:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        firstName:
          type: string
          maxLength: 50
          description: User first name
          example: John
        lastName:
          type: string
          maxLength: 50
          description: User last name
          example: Doe
        phone:
          type: string
          description: User phone number
          example: +1234567890
        role:
          type: string
          enum:
            - admin
            - user
            - driver
            - manager
          description: User role
          example: user
        status:
          type: string
          enum:
            - active
            - inactive
            - suspended
          description: User status
          example: active
        avatar:
          type: string
          description: User avatar URL
          example: https://example.com/avatar.jpg
        address:
          type: string
          description: User address
          example: 123 Main St
        city:
          type: string
          description: User city
          example: New York
        country:
          type: string
          description: User country
          example: United States
        postalCode:
          type: string
          description: User postal code
          example: 10001
        dateOfBirth:
          type: string
          format: date
          description: User date of birth
          example: '1990-01-01'
    ChangePasswordDto:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          minLength: 6
          description: Current user password
          example: currentPassword123
        newPassword:
          type: string
          minLength: 6
          description: New user password
          example: newPassword123
...